requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.2.5
Flask-Login==0.6.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.2.3
gunicorn
Flask-SQLAlchemy
python-decouple
